LayoutTests/fast/js/dfg-int32-to-double-on-known-number-expected.txt

Tests that we don't emit unnecessary speculation checks when performing an int32 to double conversion on a value that is proved to be a number, predicted to be an int32, but not proved to be an int32.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS foo({f:5}, 42.5) is 52.5
PASS successfullyParsed is true

TEST COMPLETE
